chore: вывел несколько popup в отдельные компоненты
