+++ src/components/DynamicLayout/DynamicLayout.tsx
@@ -2,7 +2,7 @@ import { useQueryClient } from '@tanstack/react-query'
 import Cookies from 'js-cookie'
 import { useRouter } from 'next/router'
 import { type FC, HTMLAttributes, PropsWithChildren, useCallback } from 'react'
-import PublicHeader from 'src/components/PublicHeader/PublicHeader'
+import PublicHeader from '@/src/components/DynamicLayout/PublicHeader/PublicHeader'
 
 import { COOKIES } from '@/src/constants/constants'
