+++ src/components/DynamicLayout/PopupCatalog/_types.ts
@@ -1,0 +1,15 @@
+export type GlobalCategoryArrayTypes = {
+  /** ссылка на страницу категории */
+  href: string
+  /** id  строки , для key */
+  id: number
+  /** иконка, для глобальной категории  */
+  img: string
+  /** название строки */
+  name: string
+}
+
+export type CatalogPopupTypes = {
+  /** состояние popup открыт/закрыт  */
+  popupOpen: boolean
+}
