+++ src/components/DynamicLayout/PopupCatalog/constants.ts
@@ -10,32 +10,6 @@ import {
   IconWomansHead
 } from '@/src/constants/icons'
 
-export const moreArray = [
-  { href: '/', id: 1, name: 'Каталог' },
-  { href: '/', id: 2, name: 'Бьюти-академия' },
-  { href: '/', id: 3, name: 'Журнал' },
-  { href: '/', id: 4, name: 'Медитация' },
-  { href: '/', id: 5, name: 'Контакты' },
-  { href: '/', id: 6, name: 'Мобильное приложение' },
-  { href: '/', id: 7, name: 'Конкурсная программа' },
-  { href: '/', id: 8, name: 'Сервисный центр' },
-  { href: '/', id: 9, name: 'Trade-In' },
-  { href: '/', id: 10, name: 'Подарочные карты' },
-  { href: '/', id: 11, name: 'Часто задаваемые вопросы' },
-  { href: '/', id: 12, name: 'Отзывы о продукции' },
-  { href: '/', id: 13, name: 'Видео отзывы' },
-  { href: '/', id: 14, name: 'Рассрочка' }
-]
-
-export const navigationLinksText = [
-  { href: '/', id: 1, name: 'Акции' },
-  { href: '/', id: 2, name: 'Магазины' },
-  { href: '/', id: 3, name: 'Виртуальный шоурум' },
-  { href: '/', id: 4, name: 'Оплата' },
-  { href: '/', id: 5, name: 'Доставка' },
-  { href: '/', id: 6, name: 'О компании' }
-]
-
 export const globalCategory = [
   { href: '/', id: 1, img: IconMassageChair, name: 'Массажные кресла', src: '/assets/images/catalogPopup/globalCategory/chair.webp' },
   { href: '/', id: 2, img: IconFoot, name: 'Массажёры для ног', src: '/assets/images/catalogPopup/globalCategory/foot.webp' },
@@ -106,88 +80,4 @@ export const subcategoriesArray = [
     { href: '/', id: 5, name: 'Валики и аксессуары' }
   ]
 
-]
-
-export const arrayCity = [
-  'Аксай',
-  'Анапа',
-  'Ангарск',
-  'Армавир',
-  'Архангельск',
-  'Астрахань',
-  'Барнаул',
-  'Белгород',
-  'Белокуриха',
-  'Березники',
-  'Бийск',
-  'Брянск',
-  'Владивосток',
-  'Владимир',
-  'Волгоград',
-  'Воронеж',
-  'Дзержинск',
-  'Евпатория',
-  'Екатеринбург',
-  'Ижевск',
-  'Иркутск',
-  'Иваново',
-  'Йошкар-Ола',
-  'Казань',
-  'Калининград',
-  'Кемерово',
-  'Кисловодск',
-  'Кострома',
-  'Краснодар',
-  'Красноярск',
-  'Курск',
-  'Липецк',
-  'Москва',
-  'Магнитогорск',
-  'Махачкала',
-  'Минск',
-  'Мурманск',
-  'Набережные Челны',
-  'Нальчик',
-  'Нижневартовск',
-  'Нижнекамск',
-  'Нижний Новгород',
-  'Новокузнецк',
-  'Новороссийск',
-  'Новосибирск',
-  'Одесса',
-  'Омск',
-  'Оренбург',
-  'Орск',
-  'Пенза',
-  'Пермь',
-  'Петропавловск-Камчатский',
-  'Псков',
-  'Пятигорск',
-  'Ростов-на-Дону',
-  'Рязань',
-  'Санкт-Петербург',
-  'Самара',
-  'Саратов',
-  'Севастополь',
-  'Северск',
-  'Симферополь',
-  'Смоленск',
-  'Сочи',
-  'Ставрополь',
-  'Сургут',
-  'Таганрог',
-  'Тверь',
-  'Тольятти',
-  'Томск',
-  'Тула',
-  'Тюмень',
-  'Улан-Удэ',
-  'Уфа',
-  'Хабаровск',
-  'Челябинск',
-  'Череповец',
-  'Южно-Сахалинск',
-  'Якутск',
-  'Ялта',
-  'Ярославль'
 ]
