+++ src/components/DynamicLayout/PopupSupport/PopupSupport.module.scss
@@ -1,0 +1,206 @@
+@import "styles/variables";
+
+.popupSupport {
+  top: 100%;
+  right: -220px;
+  z-index: 333;
+  padding-top: 15px;
+  position: absolute;
+  z-index: 444;
+  transition: all 0.2s ease;
+
+  &__wrapper {
+    display: flex;
+    width: 585px;
+    background-color: $white;
+    border-radius: 15px;
+    height: 410px;
+    padding: 14px 13px;
+    column-gap: 13px;
+
+  }
+
+  &_close {
+    visibility: hidden;
+    opacity: 0;
+    z-index: -1;
+    // transform: translateX(100%);
+    z-index: -1;
+  }
+
+  &__box {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    border-radius: 15px;
+    height: 382px;
+
+    &_left {
+      padding: 13px 10px 18px;
+      width: 298px;
+      background-color: $bad-white;
+    }
+
+    &_right {
+      padding: 13px 15px 23px;
+      width: 248px;
+      background-color: $black;
+    }
+  }
+
+  &__titleBox {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-bottom: 12px;
+  }
+
+  &__iconSupportWrapper {
+    width: 12px;
+    height: 16px;
+    min-width: unset;
+    min-height: unset;
+    margin-right: 8px;
+
+    svg {
+      width: 100%;
+      fill: $red-main;
+    }
+  }
+
+  &__title {
+    color: $black;
+    font-size: 18px;
+    font-weight: 600;
+    line-height: 160%;
+
+    font-size: 20.95px;
+
+    &_support {
+      color: $black;
+      font-weight: 700;
+    }
+
+    &_center {
+      font-weight: 400;
+      color: $white;
+    }
+  }
+
+  &__boxNumber {
+    display: flex;
+    align-items: center;
+    column-gap: 10px;
+
+    &_top {
+      margin-bottom: 5px;
+    }
+
+    &_bottom {
+      margin-bottom: 15px;
+    }
+  }
+
+  &__number {
+    color: $black;
+    font-size: 22.194px;
+    font-weight: 700;
+
+    &_white {
+      width: 100%;
+      color: $white;
+      margin-bottom: 25px;
+    }
+  }
+
+  &__text {
+    color: $black;
+    font-size: 12.684px;
+    font-weight: 400;
+    line-height: 125%;
+
+    &_underline {
+      text-decoration: underline;
+    }
+
+    &_white {
+      color: $white;
+    }
+  }
+
+  &__subtitle {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 2px;
+    width: 136px;
+    height: 25px;
+    margin-bottom: 24px;
+    padding: 4px;
+    background-color: $gray-dark;
+
+    color: $white;
+    font-size: 12.684px;
+    font-weight: 700;
+    text-transform: uppercase;
+  }
+
+  &__popupBtn {
+    width: 100%;
+    height: 50px;
+
+    &:hover {
+      svg {
+        transform: unset;
+      }
+    }
+  }
+
+  &__callBack {
+    margin-bottom: 12px;
+  }
+
+  &__videoConsultation {
+    margin-bottom: 20px;
+  }
+
+  &__iconCenterWrapper {
+    width: 20px;
+    height: 15px;
+    margin-right: 12px;
+
+    svg {
+      fill: $white;
+    }
+  }
+
+  &__listWorkingHours {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    row-gap: 8px;
+    margin-bottom: 25px;
+  }
+
+  &__itemWorkingHours {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    column-gap: 30px;
+  }
+
+  &__itemText {
+    color: $white;
+    font-size: 18px;
+    font-weight: 400;
+
+    &_day {
+      width: 40px;
+    }
+
+    &_weight {
+      font-weight: 700;
+    }
+  }
+}
