+++ src/components/DynamicLayout/PopupSupport/_types.ts
@@ -1,0 +1,4 @@
+export type PopupSupportTypes = {
+  /** состояние popup открыт/закрыт */
+  popupOpen: boolean
+}
