+++ src/components/DynamicLayout/PopupYourCity/_types.ts
@@ -1,0 +1,4 @@
+export type PopupYourCityTypes = {
+  /** состояние popup открыт/закрыт */
+  popupOpen: boolean
+}
