+++ src/components/DynamicLayout/PopupYourCity/constants.ts
@@ -1,0 +1,83 @@
+export const arrayCity = [
+  'Аксай',
+  'Анапа',
+  'Ангарск',
+  'Армавир',
+  'Архангельск',
+  'Астрахань',
+  'Барнаул',
+  'Белгород',
+  'Белокуриха',
+  'Березники',
+  'Бийск',
+  'Брянск',
+  'Владивосток',
+  'Владимир',
+  'Волгоград',
+  'Воронеж',
+  'Дзержинск',
+  'Евпатория',
+  'Екатеринбург',
+  'Ижевск',
+  'Иркутск',
+  'Иваново',
+  'Йошкар-Ола',
+  'Казань',
+  'Калининград',
+  'Кемерово',
+  'Кисловодск',
+  'Кострома',
+  'Краснодар',
+  'Красноярск',
+  'Курск',
+  'Липецк',
+  'Москва',
+  'Магнитогорск',
+  'Махачкала',
+  'Минск',
+  'Мурманск',
+  'Набережные Челны',
+  'Нальчик',
+  'Нижневартовск',
+  'Нижнекамск',
+  'Нижний Новгород',
+  'Новокузнецк',
+  'Новороссийск',
+  'Новосибирск',
+  'Одесса',
+  'Омск',
+  'Оренбург',
+  'Орск',
+  'Пенза',
+  'Пермь',
+  'Петропавловск-Камчатский',
+  'Псков',
+  'Пятигорск',
+  'Ростов-на-Дону',
+  'Рязань',
+  'Санкт-Петербург',
+  'Самара',
+  'Саратов',
+  'Севастополь',
+  'Северск',
+  'Симферополь',
+  'Смоленск',
+  'Сочи',
+  'Ставрополь',
+  'Сургут',
+  'Таганрог',
+  'Тверь',
+  'Тольятти',
+  'Томск',
+  'Тула',
+  'Тюмень',
+  'Улан-Удэ',
+  'Уфа',
+  'Хабаровск',
+  'Челябинск',
+  'Череповец',
+  'Южно-Сахалинск',
+  'Якутск',
+  'Ялта',
+  'Ярославль'
+]
