+++ src/components/DynamicLayout/PublicHeader/_types.ts
@@ -16,17 +16,6 @@ export type navLinkArrayTypes = {
     name: string
 }
 
-export type GlobalCategoryArrayTypes = {
-    /** ссылка на страницу категории */
-    href: string
-    /** id  строки , для key */
-    id: number
-    /** иконка, для глобальной категории  */
-    img: string
-    /** название строки */
-    name: string
-}
-
 export type PublicHeaderTypes = {
     /** id  строки , для key */
     id: number
