+++ src/components/DynamicLayout/PublicHeader/constants.ts
@@ -1,0 +1,25 @@
+export const moreArray = [
+  { href: '/', id: 1, name: 'Каталог' },
+  { href: '/', id: 2, name: 'Бьюти-академия' },
+  { href: '/', id: 3, name: 'Журнал' },
+  { href: '/', id: 4, name: 'Медитация' },
+  { href: '/', id: 5, name: 'Контакты' },
+  { href: '/', id: 6, name: 'Мобильное приложение' },
+  { href: '/', id: 7, name: 'Конкурсная программа' },
+  { href: '/', id: 8, name: 'Сервисный центр' },
+  { href: '/', id: 9, name: 'Trade-In' },
+  { href: '/', id: 10, name: 'Подарочные карты' },
+  { href: '/', id: 11, name: 'Часто задаваемые вопросы' },
+  { href: '/', id: 12, name: 'Отзывы о продукции' },
+  { href: '/', id: 13, name: 'Видео отзывы' },
+  { href: '/', id: 14, name: 'Рассрочка' }
+]
+
+export const navigationLinksText = [
+  { href: '/', id: 1, name: 'Акции' },
+  { href: '/', id: 2, name: 'Магазины' },
+  { href: '/', id: 3, name: 'Виртуальный шоурум' },
+  { href: '/', id: 4, name: 'Оплата' },
+  { href: '/', id: 5, name: 'Доставка' },
+  { href: '/', id: 6, name: 'О компании' }
+]
